python-worker/requirements.txt
google-genai==1.34.0
requests==2.32.3
fastapi>=0.115.0
pydantic>=2.0.0
supabase>=2.7.0
